image-processing-queue/api/requirements.txt
fastapi
uvicorn[standard]
minio
pika
Pillow
python-multipart
